# Conflicts:
#	src/main/java/com/book/management/book/config/AppConfig.java
#	src/main/java/com/book/management/book/controller/BookController.java
#	src/main/java/com/book/management/book/exception/GlobalBookExceptionHandler.java
#	src/main/java/com/book/management/book/exception/GlobalExceptionHandler.java
#	src/main/java/com/book/management/book/repository/BookRepository.java
#	src/main/java/com/book/management/book/repository/impl/InMemoryBookRepository.java
#	src/main/java/com/book/management/book/service/BookService.java
#	src/main/java/com/book/management/book/service/impl/BookServiceImpl.java
#	src/main/java/com/book/management/review_rating/config/AppConfig.java
#	src/main/java/com/db/ms/book/config/AppConfig.java
#	src/main/java/com/db/ms/book/exception/GlobalExceptionHandler.java

+++ src/main/java/com/book/management/book/dto/requestdto/AddBookRequestDTO.java
@@ -13,7 +13,7 @@ public class AddBookRequestDTO {
     private String bookAuthorId;
     private String bookCategoryId;     // e.g., "CAT-FIC"
     private Double bookPrice;          // non-negative
-    private Long bookStockQuantity;    // non-negative
+    private Integer bookStockQuantity;    // non-negative
 
     public Long getBookId() {
         return bookId;
@@ -35,7 +35,7 @@ public class AddBookRequestDTO {
         return bookPrice;
     }
 
-    public Long getBookStockQuantity() {
+    public Integer getBookStockQuantity() {
         return bookStockQuantity;
     }
 }
